requirements.txt
appnope==0.1.0
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
blockstack-zones-py3==0.19.0.0
certifi==2020.6.20
chardet==3.0.4
colorlog==4.4.0
decorator==4.4.2
deepdiff==5.0.2
Django==3.1
django-environ==0.4.5
django-filter==2.3.0
django-mysql==3.8.1
django-yamlfield==1.0.3
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
ipaddr==2.2.0
ipython-genutils==0.2.0
isort==5.5.1
jedi==0.17.2
Jinja2==3.0.0a1
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==2.0.0a1
marshmallow==3.8.0
mccabe==0.6.1
mysqlclient==2.0.1
ordered-set==4.0.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
setproctitle==1.1.10
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.1.0
